ORC-418. Fix broken centos6 compilation.
Fixes #328

+++ c++/src/RLEv1.cc
@@ -25,14 +25,14 @@
 
 namespace orc {
 
-const int MINIMUM_REPEAT = 3;
-const int MAXIMUM_REPEAT = 127 + MINIMUM_REPEAT;
+const uint64_t MINIMUM_REPEAT = 3;
+const uint64_t MAXIMUM_REPEAT = 127 + MINIMUM_REPEAT;
 
 const int64_t BASE_128_MASK = 0x7f;
 
-const int MAX_DELTA = 127;
-const int MIN_DELTA = -128;
-const int MAX_LITERAL_SIZE = 128;
+const int64_t MAX_DELTA = 127;
+const int64_t MIN_DELTA = -128;
+const uint64_t MAX_LITERAL_SIZE = 128;
 
 RleEncoderV1::RleEncoderV1(
                           std::unique_ptr<BufferedOutputStream> outStream,
